add procedure to get all deadline for a school

+++ Database Schema/HorizanDBProcedures.sql
@@ -958,8 +958,8 @@ End
  */
 
 --  MARK: update school summary. need deadline cycle and student type info
-drop Procedure pGetSchoolSummary;
 Use HorizanDB;
+drop Procedure pGetSchoolSummary;
 Create
 Procedure pGetSchoolSummary(
     User_First_Name VarChar(20),
@@ -1172,7 +1172,7 @@ possible data:
  
 Use HorizanDB;
 Create
-Procedure pGetSchoolDeadline(
+Procedure pGetSchoolDeadlineInfo(
     School_Name VarChar(255),
     Deadline_Name Varchar(50),
     DeadlineCycle_Year int,
@@ -1226,6 +1226,38 @@ begin
             and sd.DeadlineID = Deadline_ID
             and sd.DeadlineCycleID = DeadlineCycle_ID
             and sd.StudentTypeID = StudentType_ID
+        );
+        Select * from tempDeadlineCycle;
+        drop table tempDeadlineCycle;
+    end if;
+end;
+
+
+Use HorizanDB;
+drop procedure pGetAllDeadlines;
+Create
+Procedure pGetAllDeadlines(
+    School_Name VarChar(255)
+)
+begin 
+    Declare School_ID int;
+
+    Call pGetSchool (School_Name, School_ID);
+    if School_ID is null
+    then
+        SIGNAL SQLSTATE '45000'
+        Set MESSAGE_TEXT = 'School not found';
+    end if;
+
+    if @@error_count = 0
+    then 
+        create temporary table tempDeadlineCycle(
+            Select  dd.DeadlineName, sd.DeadlineDatetime,dc.DeadlineCycleYear, dc.DeadlineCycleSeason, st.StudentTypeName
+            from SchoolDeadline sd
+            join DeadlineCycle dc on sd.DeadlineCycleID = dc.DeadlineCycleID
+            join DeadlineDetail dd on sd.DeadlineID = dd.DeadlineID
+            join StudentType st on sd.StudentTypeID = st.StudentTypeID
+            where sd.SchoolID = School_ID
         );
         Select * from tempDeadlineCycle;
         drop table tempDeadlineCycle;
